Tests/LibWeb/Text/input/HTML/Navigation-object-properties.html

<script src="../include.js"></script>
<script>
    test(() => {
        let n = window.navigation;

        let len = n.entries().length;

        println(`entries[length - 1] is current entry: ${n.entries()[len - 1] === n.currentEntry}`);
        println(`currentEntry is a ${n.currentEntry}`);

        println(`transition is null: ${n.transition == null}`);
        println(`canGoBack: ${n.canGoBack}`);
        println(`canGoForward: ${n.canGoForward}`);       
    });
</script>
