Tests/LibWeb/Text/expected/HTML/Navigation-object-properties.txt

entries[length - 1] is current entry: true
currentEntry is a [object NavigationHistoryEntry]
transition is null: true
canGoBack: true
canGoForward: true
